cleaned the code

+++ commerce/models.py
@@ -13,9 +13,10 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price =  models.FloatField()
+    price =  models.DecimalField(max_digits=9, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=False)
     image = models.ImageField(null=True, blank=True)
+
     def __str__(self):
         return self.name
